asf properly added to user file
Base commit as there is a lot of work here that is unclear.

The ASF project is very important to this project. It's also huge, and not designed for IAR, it was written for Atmel Studio before being updated to be used in the Microchip equivilent. It comes with a wizard that should be easy to configure, however although built in to the Microchip IDE it is bug ridden and considered beta. In theory you can generate a makefile for IAR but no-one seems to have maked it work, and it's unsupported. There is plenty of documentation online on using it, but not configuring it. Also IAR requires it to be used as a separate project linked to the main project, which is difficult and involes more tools and reasonable compliler knowledge, which can't be trimmed in IAR easily.

As you cannot add file structures to IAR (it only shows folders for user ease, the files are held in XML in the project file) it takes around 4 hours to manually add groups and drag files into deep structures, and many files will not compile when used together as they use different files named the same but with different defines.

The wizard does not support our processor anyway, and the makefile it produces for supported processors are poorly presented.

This base concept is hand crafted, recognises our processor correctly, and compiles fine. I've added two drivers to the ASF, the compilation information and utility support. Every further driver added will require hand adding the required headers, although this gets increasingly easy as the correct support files tend to be previously added. Adding high level modules should be much easier now.

Anyone wishing to add to to it should add a group at a time, add it manually to the IDE file viewer and compile fixing all linking issues before testing carefully that you're not using two conflicting headers (IAR is unable to identify files with the same name and gives random results if driven from the IDE).